ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
